refactor(ts): separate different types of providers
